chore: file architecture reorganization
